Remove concept of orphan node as it can put the graph into corrupt state
This commit removes the detection of orphan nodes, which triggered the removal of unused selectors. The problem
is that the selector could be used somewhere else in the graph, even if this node doesn't have any incoming edge.
Basically it means that the strategy for "eliminating previous constraints" when a subgraph is elimintated doesn't
work. As a consequence some tests are now marked as `@NotYetImplemented`.